End of section 5

+++ src/AuctionService/Controller/AuctionsController.cs
@@ -5,6 +5,7 @@ using AutoMapper;
 using AutoMapper.QueryableExtensions;
 using Contracts;
 using MassTransit;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -51,12 +52,13 @@ namespace AuctionService.Controller
             return _mapper.Map<AuctionDto>(auction);
         }
 
+        [Authorize]
         [HttpPost]
         public async Task<ActionResult<AuctionDto>> CreateAuction(CreateAuctionDto auctionDto)
         {
             var auction = _mapper.Map<Auction>(auctionDto);
 
-            auction.Seller = "test";
+            auction.Seller = User!.Identity!.Name;
 
             _context.Auctions.Add(auction);
 
@@ -71,6 +73,7 @@ namespace AuctionService.Controller
             return CreatedAtAction(nameof(GetAuctionById), new { auction.Id }, newAuction);
         }
 
+        [Authorize]
         [HttpPut("{id}")]
         public async Task<ActionResult> UpdateAuction(Guid id, UpdateAuctionDto updateAuctionDto)
         {
@@ -80,6 +83,8 @@ namespace AuctionService.Controller
                 .FirstOrDefaultAsync(x => x.Id == id);
 
             if (auction == null) return NotFound();
+
+            if (auction.Seller != User!.Identity!.Name) return Forbid();
 
             auction.Item.Make = updateAuctionDto.Make ?? auction.Item.Make;
             auction.Item.Model = updateAuctionDto.Model ?? auction.Item.Model;
@@ -96,6 +101,7 @@ namespace AuctionService.Controller
             return BadRequest("Problem saving changes");
         }
 
+        [Authorize]
         [HttpDelete("{id}")]
         public async Task<ActionResult> DeleteAuction(Guid id)
         {
@@ -105,6 +111,8 @@ namespace AuctionService.Controller
                 .FirstOrDefaultAsync(x => x.Id == id);
 
             if (auction == null) return NotFound();
+
+            if (auction.Seller != User!.Identity!.Name) return Forbid();
 
             _context.Auctions.Remove(auction);
